+++ server/src-lib/Hasura/Backends/MSSQL/Plan.hs
@@ -13,7 +13,6 @@ import qualified Data.Aeson                    as J
 import qualified Data.HashMap.Strict           as HM
 import qualified Data.HashMap.Strict.InsOrd    as OMap
 import qualified Data.HashSet                  as Set
-import qualified Data.List.NonEmpty            as NE
 import qualified Data.Text                     as T
 import qualified Database.ODBC.SQLServer       as ODBC
 import qualified Language.GraphQL.Draft.Syntax as G
@@ -59,21 +58,22 @@ planNoPlan userInfo queryDB = do
             JsonFor forJson -> JsonFor forJson {jsonRoot = Root "root"}
       }
 
-  -- planMultiplex ::
-  --      OMap.InsOrdHashMap G.Name (SubscriptionRootFieldMSSQL (GraphQL.UnpreparedValue 'MSSQL))
-  --   -> Either PrepareError Select
-  -- planMultiplex _unpreparedMap =
-  -- let rootFieldMap =
-  --       evalState
-  --         (traverse
-  --            (traverseQueryRootField prepareValueMultiplex)
-  --            unpreparedMap)
-  --         emptyPrepareState
-  -- selectMap <-
-  --   first
-  --     FromIrError
-  --     (runValidate (TSQL.runFromIr (traverse TSQL.fromRootField rootFieldMap)))
-  -- pure (multiplexRootReselect (collapseMap selectMap))
+planMultiplex
+  :: MonadError QErr m
+  => OMap.InsOrdHashMap G.Name (QueryDB 'MSSQL (GraphQL.UnpreparedValue 'MSSQL))
+  -> SessionVariables
+  -> m (Reselect, PrepareState)
+planMultiplex unpreparedMap sessionVariables = do
+  let (rootFieldMap, prepareState) =
+        runState
+          (traverse
+            (traverseQueryDB (prepareValueMultiplex (getSessionVariablesSet sessionVariables)))
+            unpreparedMap)
+          emptyPrepareState
+  selectMap <-
+    runValidate (TSQL.runFromIr (traverse TSQL.fromRootField rootFieldMap))
+    `onLeft` (throw400 NotSupported . tshow)
+  pure (collapseMap selectMap, prepareState)
 
   -- Plan a query without prepare/exec.
   -- planNoPlanMap ::
@@ -119,10 +119,6 @@ globalSessionExpression :: TSQL.Expression
 globalSessionExpression =
   ValueExpression (ODBC.TextValue "current_setting('hasura.user')::json")
 
--- TODO: real env object.
-envObjectExpression :: TSQL.Expression
-envObjectExpression =
-  ValueExpression (ODBC.TextValue "[{\"result_id\":1,\"result_vars\":{\"synthetic\":[10]}}]")
 
 --------------------------------------------------------------------------------
 -- Resolving values
@@ -131,14 +127,17 @@ data PrepareError
   = FromIrError (NonEmpty TSQL.Error)
 
 data PrepareState = PrepareState
-  { positionalArguments :: !Integer
+  { positionalArguments :: ![RQL.ColumnValue 'MSSQL]
   , namedArguments      :: !(HashMap G.Name (RQL.ColumnValue 'MSSQL))
   , sessionVariables    :: !(Set.HashSet SessionVariable)
   }
 
 emptyPrepareState :: PrepareState
-emptyPrepareState =
-  PrepareState {positionalArguments = 0, namedArguments = mempty, sessionVariables = mempty}
+emptyPrepareState = PrepareState
+  { positionalArguments = mempty
+  , namedArguments = mempty
+  , sessionVariables = mempty
+  }
 
 -- | Prepare a value without any query planning; we just execute the
 -- query with the values embedded.
@@ -158,22 +157,37 @@ prepareValueNoPlan sessionVariables =
       pure $ ValueExpression $ ODBC.TextValue value
 
 -- | Prepare a value for multiplexed queries.
-prepareValueMultiplex ::
-     GraphQL.UnpreparedValue 'MSSQL
+prepareValueMultiplex
+  :: Set.HashSet SessionVariable
+  -> GraphQL.UnpreparedValue 'MSSQL
   -> State PrepareState TSQL.Expression
-prepareValueMultiplex =
+prepareValueMultiplex globalVariables =
   \case
     GraphQL.UVLiteral x -> pure x
-    GraphQL.UVSession ->
-      pure (JsonQueryExpression globalSessionExpression)
+    GraphQL.UVSession -> do
+      modify' (\s -> s {sessionVariables = sessionVariables s <> globalVariables})
+      pure $ JsonValueExpression
+        (ColumnExpression
+          FieldName
+          { fieldNameEntity = rowAlias
+          , fieldName = resultVarsAlias
+          })
+        (RootPath `FieldPath` "session")
     GraphQL.UVSessionVar _typ text -> do
       modify' (\s -> s {sessionVariables = text `Set.insert` sessionVariables s})
-      pure $ JsonValueExpression globalSessionExpression (FieldPath RootPath (toTxt text))
-    GraphQL.UVParameter mVariableInfo pgcolumnvalue ->
+      pure $ JsonValueExpression
+        (ColumnExpression
+          FieldName
+          { fieldNameEntity = rowAlias
+          , fieldName = resultVarsAlias
+          })
+        (RootPath `FieldPath` "session" `FieldPath` toTxt text)
+    GraphQL.UVParameter mVariableInfo columnValue ->
       case fmap GraphQL.getName mVariableInfo of
         Nothing -> do
-          index <- gets positionalArguments
-          modify' (\s -> s {positionalArguments = index + 1})
+          modify' (\s -> s {
+            positionalArguments = positionalArguments s <> [columnValue] })
+          index <- (toInteger . length) <$> gets positionalArguments
           pure
             (JsonValueExpression
                (ColumnExpression
@@ -187,76 +201,17 @@ prepareValueMultiplex =
             (\s ->
                s
                  { namedArguments =
-                     HM.insert name pgcolumnvalue (namedArguments s)
+                     HM.insert name columnValue (namedArguments s)
                  })
           pure
             (JsonValueExpression
-               envObjectExpression
+               (ColumnExpression
+                  FieldName
+                    { fieldNameEntity = rowAlias
+                    , fieldName = resultVarsAlias
+                    })
                (RootPath `FieldPath` "query" `FieldPath` G.unName name))
 
---------------------------------------------------------------------------------
--- Producing the correct SQL-level list comprehension to multiplex a query
-
--- Problem description:
---
--- Generate a query that repeats the same query N times but with
--- certain slots replaced:
---
--- [ Select x y | (x,y) <- [..] ]
---
-
-multiplexRootReselect :: TSQL.Reselect -> TSQL.Select
-multiplexRootReselect rootReselect =
-  Select
-    { selectTop = NoTop
-    , selectProjections =
-        [ FieldNameProjection
-            Aliased
-              { aliasedThing =
-                  FieldName
-                    {fieldNameEntity = rowAlias, fieldName = resultIdAlias}
-              , aliasedAlias = resultIdAlias
-              }
-        , ExpressionProjection
-            Aliased
-              { aliasedThing =
-                  JsonQueryExpression
-                    (ColumnExpression
-                       (FieldName
-                          { fieldNameEntity = resultAlias
-                          , fieldName = TSQL.jsonFieldName
-                          }))
-              , aliasedAlias = resultAlias
-              }
-        ]
-    , selectFrom =
-        FromOpenJson
-          Aliased
-            { aliasedThing =
-                OpenJson
-                  { openJsonExpression = envObjectExpression
-                  , openJsonWith =
-                      NE.fromList
-                        [IntField resultIdAlias, JsonField resultVarsAlias]
-                  }
-            , aliasedAlias = rowAlias
-            }
-    , selectJoins =
-        [ Join
-            { joinSource = JoinReselect rootReselect
-            , joinJoinAlias =
-                JoinAlias
-                  { joinAliasEntity = resultAlias
-                  , joinAliasField = Just TSQL.jsonFieldName
-                  }
-            }
-        ]
-    , selectWhere = Where mempty
-    , selectFor =
-        JsonFor ForJson {jsonCardinality = JsonArray, jsonRoot = NoRoot}
-    , selectOrderBy = Nothing
-    , selectOffset = Nothing
-    }
 
 resultIdAlias :: T.Text
 resultIdAlias = "result_id"