+++ server/src-lib/Hasura/GraphQL/Execute/LiveQuery/Plan.hs
@@ -125,7 +125,7 @@ import           Hasura.Session
 -- Cohort
 
 newtype CohortId = CohortId { unCohortId :: UUID }
-  deriving (Show, Eq, Hashable, J.ToJSON, Q.FromCol)
+  deriving (Show, Eq, Hashable, J.ToJSON, J.FromJSON, Q.FromCol)
 
 newCohortId :: (MonadIO m) => m CohortId
 newCohortId = CohortId <$> liftIO UUID.nextRandom
@@ -233,14 +233,14 @@ instance Q.ToPrepArg CohortVariablesArray where
 --
 -- so if any variable values are invalid, the error will be caught early.
 
-newtype ValidatedVariables f = ValidatedVariables (f TxtEncodedPGVal)
+newtype ValidatedVariables f = ValidatedVariables (f TxtEncodedVal)
 
-deriving instance (Show      (f TxtEncodedPGVal)) => Show      (ValidatedVariables f)
-deriving instance (Eq        (f TxtEncodedPGVal)) => Eq        (ValidatedVariables f)
-deriving instance (Hashable  (f TxtEncodedPGVal)) => Hashable  (ValidatedVariables f)
-deriving instance (J.ToJSON  (f TxtEncodedPGVal)) => J.ToJSON  (ValidatedVariables f)
-deriving instance (Semigroup (f TxtEncodedPGVal)) => Semigroup (ValidatedVariables f)
-deriving instance (Monoid    (f TxtEncodedPGVal)) => Monoid    (ValidatedVariables f)
+deriving instance (Show (f TxtEncodedVal)) => Show (ValidatedVariables f)
+deriving instance (Eq (f TxtEncodedVal)) => Eq (ValidatedVariables f)
+deriving instance (Hashable (f TxtEncodedVal)) => Hashable (ValidatedVariables f)
+deriving instance (J.ToJSON (f TxtEncodedVal)) => J.ToJSON (ValidatedVariables f)
+deriving instance (Semigroup (f TxtEncodedVal)) => Semigroup (ValidatedVariables f)
+deriving instance (Monoid (f TxtEncodedVal)) => Monoid (ValidatedVariables f)
 
 type ValidatedQueryVariables     = ValidatedVariables (Map.HashMap G.Name)
 type ValidatedSyntheticVariables = ValidatedVariables []